requirements.txt
Django==5.1.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.1.0
psycopg2-binary==2.9.9
celery==5.3.4
djangorestframework-simplejwt==5.3.0